✅ feat: Add test for Apollo
TODO: implement Apollo mock
