+++ Targets/MorningBearKit/Tests/Network/RequestServiceTests.swift
@@ -15,54 +15,7 @@ import RxSwift
 @testable import MorningBearKit
 
 final class RequestServiceTests: XCTestCase {
-    let mockData = FakeData(name: "test")
-    let bag = DisposeBag()
     
-    func test__request_get_200() throws {
-        let mockProvider = stubProvider(endpointClosure: mBearEndpointClosure(200))
-        let expectation = XCTestExpectation(description: "Network")
-        
-        MBearAPI.login(load: mockData).request(provider: mockProvider)
-            .map { response -> FakeData in
-                let data = response.data
-                
-                print(String(data: data, encoding: .utf8) as Any)
-                let decoded = try MBearAPI.jsonDecoder.decode(FakeData.self, from: data)
-                return decoded
-            }
-            .subscribe(
-                onSuccess: { post in
-                    expectation.fulfill()
-                }, onFailure: { error in
-                    XCTFail("Test get failed: \(error)")
-                })
-            .disposed(by: bag)
-        
-        wait(for: [expectation], timeout: 3)
-    }
-    
-    func test__request_get_400() throws {
-        let expectation = XCTestExpectation(description: "Network")
-        let errorStubProvider = stubProvider(endpointClosure: mBearEndpointClosure(400))
-        
-        MBearAPI.login(load: mockData).request(provider: errorStubProvider)
-            .map { response -> FakeData in
-                let data = response.data
-                
-                print(String(data: data, encoding: .utf8) as Any)
-                let decoded = try MBearAPI.jsonDecoder.decode(FakeData.self, from: data)
-                return decoded
-            }
-            .subscribe(
-                onSuccess: { post in
-                    XCTFail("Test get failed: \(post)")
-                }, onFailure: { error in
-                    expectation.fulfill()
-                })
-            .disposed(by: bag)
-        
-        wait(for: [expectation], timeout: 3)
-    }
 }
 
 // MARK: - Internal testing tools
